use uint64_t instead of uint32_t

+++ lab02/test.c
@@ -1,16 +1,16 @@
 #include <stdio.h>
 #include <stdint.h>
 
-uint32_t hex(uint32_t value) {
-  uint32_t stack[8];
-  uint32_t *p = stack;
+uint64_t hex(uint64_t value) {
+  uint64_t stack[8];
+  uint64_t *p = stack;
   while (value > 0) {
     *p = value % 16;
     value >>= 4;
     p++;
   }
   --p;
-  uint32_t result = 0;
+  uint64_t result = 0;
   while (p != stack - 1) {
     result <<= 4;
     result += *p;
@@ -20,7 +20,7 @@ uint32_t hex(uint32_t value) {
 }
 
 int main() {
-  uint32_t res = hex(200110619);
+  uint64_t res = hex(200110619);
   asm volatile (
     "addi s3, %[r], 0"
     :[r]"=r"(res)
